refactor : NearPlaceApiService 없애고, TMap으로 통일 & getNearPlaces() 함수 Query 기본 값 미리 넣어주기
