+++ data/src/main/java/com/stop/data/remote/source/nearplace/NearPlaceRemoteDataSource.kt
@@ -5,11 +5,9 @@ import com.stop.data.model.nearplace.Place
 interface NearPlaceRemoteDataSource {
 
     suspend fun getNearPlaces(
-        version: Int,
         searchKeyword: String,
         centerLon: Double,
-        centerLat: Double,
-        appKey: String
+        centerLat: Double
     ): Result<List<Place>>
 
 }
