+++ domain/src/main/java/com/stop/domain/usecase/nearplace/GetNearPlacesUseCase.kt
@@ -5,11 +5,9 @@ import com.stop.domain.model.nearplace.Place
 interface GetNearPlacesUseCase {
 
     suspend fun getNearPlaces(
-        version: Int,
         searchKeyword: String,
         centerLon: Double,
-        centerLat: Double,
-        appKey: String
+        centerLat: Double
     ): List<Place>
 
 }
