+++ domain/src/main/java/com/stop/domain/usecase/nowlocation/GetNowStationLocationUseCase.kt
@@ -5,11 +5,9 @@ import com.stop.domain.model.nowlocation.NowStationLocationUseCaseItem
 interface GetNowStationLocationUseCase {
 
     suspend operator fun invoke(
-        version: Int,
         searchKeyword: String,
         centerLon: Double,
-        centerLat: Double,
-        appKey: String
+        centerLat: Double
     ): NowStationLocationUseCaseItem
 
 }
