+++ presentation/src/main/java/com/stop/AlarmWorker.kt
@@ -7,12 +7,10 @@ import androidx.work.CoroutineWorker
 import androidx.work.Data
 import androidx.work.WorkerParameters
 import androidx.work.workDataOf
-import com.stop.domain.model.nearplace.Place
 import com.stop.domain.usecase.nearplace.GetNearPlacesUseCase
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.withContext
 
 // 일단은 NearPlaceUseCase 이용하여서 데이터 가져오기 구현
@@ -38,11 +36,9 @@ class AlarmWorker @AssistedInject constructor(
     private suspend fun callApi() {
         withContext(Dispatchers.IO) {
             val list = getNearPlacesUseCase.getNearPlaces(
-                1,
                 "아남타워",
                 126.969652,
-                37.553836,
-                BuildConfig.TMAP_APP_KEY
+                37.553836
             )
             Log.e("ABC", list.toString())
         }
