+++ presentation/src/main/java/com/stop/ui/placesearch/PlaceSearchViewModel.kt
@@ -1,14 +1,11 @@
 package com.stop.ui.placesearch
 
-import android.location.LocationManager
 import android.text.Editable
-import android.util.Log
 import android.view.View
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.stop.BuildConfig
 import com.stop.domain.model.geoLocation.GeoLocationInfo
 import com.stop.domain.model.nearplace.Place
 import com.stop.domain.usecase.geoLocation.GeoLocationUseCase
@@ -22,9 +19,7 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
-import kotlin.math.pow
 import kotlin.math.round
-import kotlin.math.sqrt
 
 @HiltViewModel
 class PlaceSearchViewModel @Inject constructor(
@@ -71,17 +66,14 @@ class PlaceSearchViewModel @Inject constructor(
             try {
                 _nearPlaceList.emit(
                     getNearPlacesUseCase.getNearPlaces(
-                        TMAP_VERSION,
                         searchKeyword,
                         currentLocation.longitude,
-                        currentLocation.latitude,
-                        BuildConfig.TMAP_APP_KEY
+                        currentLocation.latitude
                     )
                 )
             } catch (e: Exception) {
                 setNearPlaceListEmpty()
                 errorMessageChannel.send(e.message ?: "something is wrong")
-                Log.d("PlaceSearchViewModel", "getNearPlace 실패~ ${e.toString()}")
             }
         }
     }
@@ -120,7 +112,6 @@ class PlaceSearchViewModel @Inject constructor(
     }
 
     companion object {
-        private const val TMAP_VERSION = 1
         private val EXAMPLE_BOOKMARK_1 = Location(37.3931010, 126.9781449)
         private val EXAMPLE_BOOKMARK_2 = Location(37.55063543842469, 127.07369927986392)
         private val EXAMPLE_BOOKMARK_3 = Location(37.48450549635376, 126.89324337770405)
